# target/classes/openapi.yaml
openapi: 3.0.0
info:
  title: Repair Tips API
  description: API Backend para Sistema de Consulta de Dicas de Conserto de Equipamentos
  version: 1.0.0
  contact:
    name: API Support
    url: https://example.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api
    description: Development Server
  - url: https://api.example.com/api
    description: Production Server
...
paths:
...
  /pecas:
    get:
      tags:
        - Peças
      summary: Listar todas as peças
      operationId: getPecas
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de peças
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePecaResponseDTO'
    post:
      tags:
        - Peças
      summary: Criar nova peça
      operationId: createPeca
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PecaRequestDTO'
      responses:
        '201':
          description: Peça criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PecaResponseDTO'
  /pecas/{id}:
    get:
      tags:
        - Peças
      summary: Obter peça por ID
      operationId: getPecaById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Peça encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PecaResponseDTO'
    put:
      tags:
        - Peças
      summary: Atualizar peça
      operationId: updatePeca
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PecaRequestDTO'
      responses:
        '200':
          description: Peça atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PecaResponseDTO'
    delete:
      tags:
        - Peças
      summary: Deletar peça
      operationId: deletePeca
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Peça deletada com sucesso
components:
  schemas:
...
    PecaRequestDTO:
      type: object
      required:
        - nome
        - partNumber
        - categoria
        - equipamentoId
      properties:
        nome:
          type: string
          description: Nome da peça
        partNumber:
          type: string
          description: Número da parte
        categoria:
          type: string
          description: Categoria da peça
        equipamentoId:
          type: integer
          format: int64
          description: ID do equipamento
    PecaResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        partNumber:
          type: string
        categoria:
          type: string
        equipamentoId:
          type: integer
          format: int64
...
    PagePecaResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PecaResponseDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer