# target/generated-sources/openapi/src/main/java/com/repairsystem/api/PeasApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.8.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.repairsystem.api;

import com.repairsystem.dto.PagePecaResponseDTO;
import com.repairsystem.dto.PecaRequestDTO;
import com.repairsystem.dto.PecaResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-23T23:19:45.937138696-03:00[America/Sao_Paulo]", comments = "Generator version: 7.8.0")
@Validated
@Tag(name = "Peças", description = "Operações com peças de equipamentos")
public interface PeasApi {

    default PeasApiDelegate getDelegate() {
        return new PeasApiDelegate() {};
    }

    /**
     * POST /pecas : Criar nova peça
     *
     * @param pecaRequestDTO  (required)
     * @return Peça criada com sucesso (status code 201)
     */
    @Operation(
        operationId = "createPeca",
        summary = "Criar nova peça",
        tags = { "Peças" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Peça criada com sucesso", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PecaResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/pecas",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PecaResponseDTO> createPeca(
        @Parameter(name = "PecaRequestDTO", description = "", required = true) @Valid @RequestBody PecaRequestDTO pecaRequestDTO
    ) {
        return getDelegate().createPeca(pecaRequestDTO);
    }


    /**
     * DELETE /pecas/{id} : Deletar peça
     *
     * @param id  (required)
     * @return Peça deletada com sucesso (status code 204)
     */
    @Operation(
        operationId = "deletePeca",
        summary = "Deletar peça",
        tags = { "Peças" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Peça deletada com sucesso")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/pecas/{id}"
    )
    
    default ResponseEntity<Void> deletePeca(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return getDelegate().deletePeca(id);
    }


    /**
     * GET /pecas/{id} : Obter peça por ID
     *
     * @param id  (required)
     * @return Peça encontrada (status code 200)
     */
    @Operation(
        operationId = "getPecaById",
        summary = "Obter peça por ID",
        tags = { "Peças" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Peça encontrada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PecaResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/pecas/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PecaResponseDTO> getPecaById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return getDelegate().getPecaById(id);
    }


    /**
     * GET /pecas : Listar todas as peças
     *
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Lista de peças (status code 200)
     */
    @Operation(
        operationId = "getPecas",
        summary = "Listar todas as peças",
        tags = { "Peças" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Lista de peças", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PagePecaResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/pecas",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PagePecaResponseDTO> getPecas(
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Optional<Integer> page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Optional<Integer> size
    ) {
        return getDelegate().getPecas(page, size);
    }


    /**
     * PUT /pecas/{id} : Atualizar peça
     *
     * @param id  (required)
     * @param pecaRequestDTO  (required)
     * @return Peça atualizada (status code 200)
     */
    @Operation(
        operationId = "updatePeca",
        summary = "Atualizar peça",
        tags = { "Peças" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Peça atualizada", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PecaResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/pecas/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PecaResponseDTO> updatePeca(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "PecaRequestDTO", description = "", required = true) @Valid @RequestBody PecaRequestDTO pecaRequestDTO
    ) {
        return getDelegate().updatePeca(id, pecaRequestDTO);
    }

}